% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{as_points}
\alias{as_points}
\title{Convert raster object to point data}
\usage{
as_points(x = make_dummy_mask(), fun = NULL, ...)
}
\arguments{
\item{x}{raster mask as BasicRaster or SpatRaster}

\item{fun}{function or NULL, to filter points - see \code{\link[raster]{rasterToPoints}}}

\item{...}{other arguments passed through}
}
\value{
tibble of 
\itemize{
\item{x ala longitude}
\item{y ala latitude}
\item{value value at cell}
}
}
\description{
Convert raster object to point data
}
