% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asf_download.R
\name{asf_download}
\alias{asf_download}
\title{asf_download}
\usage{
asf_download(results, id, user, password)
}
\arguments{
\item{results}{A data frame retrieve from asf_search function.}

\item{id}{Row number from selected product in data frame retrieved by
asf_search function for download.}

\item{user}{Username for Alaska Satellite Facility platform.}

\item{password}{Key for Alaska Satellite Facility platform.}
}
\description{
Download selected products from Alaska Satellite Facility platform.
}
\examples{

\dontrun{
library(sf)

polygon <- cbind(c(-75.63, -75.49, -75.49, -75.63, -75.63),
                c(6.36, 6.36, 6.17, 6.17, 6.36)) \%>\%
 {st_polygon(list(.))} \%>\%
 st_sfc(crs = 4326)

results <- asf_search(data = polygon, results = 20,
 platform = 'ALOS', instrument = 'PALSAR')

asf_download(results = results, id = 12,
 user = 'your_ASF_user', password = 'your_ASF_key')
}

}
\seealso{
https://search.asf.alaska.edu/
}
