% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asf_search.R
\name{asf_search}
\alias{asf_search}
\title{asf_search}
\usage{
asf_search(data, results = 25, platform, instrument = NULL, start = NULL)
}
\arguments{
\item{data}{A sf object which represent de area where search the selected
ASF product.}

\item{results}{An integer, specifying the maximum number of results returned
from search.}

\item{platform}{A string specifying the ASF product for search. Available
products are:

ALOS, A3, AIRSAR, AS, ERS, ERS-1, E1, ERS-2, E2, JERS-1, J1, RADARSAT-1, R1,
SEASAT, SS, S1, Sentinel, Sentinel-1, Sentinel-1A, SA, Sentinel-1B, SB,
SIR-C, SMAP, SP, UAVSAR, UA}

\item{instrument}{Default NULL. For some platforms, such as ALOS, there are
multiple instruments to choose from.}

\item{start}{Start Date of search. Default NULL.}
}
\value{
A data frame with available products for the selected area.
}
\description{
Retrieve a collection with available ASF products.
}
\examples{

\dontrun{
library(sf)

polygon <- cbind(c(-75.63, -75.49, -75.49, -75.63, -75.63),
                c(6.36, 6.36, 6.17, 6.17, 6.36))
 {st_polygon(list(.))} \%>\%
 st_sfc(crs = 4326)

asf_search(data = polygon, results = 5,
           platform = 'ALOS', instrument = 'PALSAR')
}

}
\seealso{
https://docs.asf.alaska.edu/api/keywords/
}
