% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb.R
\name{bb_split}
\alias{bb_split}
\title{Split a bounding box into two at \code{at}}
\usage{
bb_split(bb = c(-170, 50, -60, 60), at = 0)
}
\arguments{
\item{bb}{numeric, 4 element bouding box of left, right, bottom and top coordinates}

\item{at}{numeric, longitude to split around}
}
\value{
list of one or two bounding box vectors
}
\description{
Split a bounding box into two at \code{at}
}
