% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb.R
\name{bb_straddles}
\alias{bb_straddles}
\title{Test if a blunding box straddles a longitude}
\usage{
bb_straddles(bb = c(-170, 50, -60, 60), at = 0)
}
\arguments{
\item{bb}{numeric, 4 element bouding box}

\item{at}{numeric, longitude to split around}
}
\value{
logical
}
\description{
Test if a blunding box straddles a longitude
}
