% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{bbox_to_matrix}
\alias{bbox_to_matrix}
\title{Convert a 4-element bbox vector to a matrix of two columns (x and y)}
\usage{
bbox_to_matrix(x = c(-72, -63, 39, 46), close = TRUE)
}
\arguments{
\item{x}{a 4-element numeric vector of left, bottom, right, top coordinates}

\item{close}{logical, if TRUE then close the polygon such that the first 
and last verices are the same}
}
\value{
a matrix of 2 columns and either 5 rows (closed) or 4 rows (open)
}
\description{
Convert a 4-element bbox vector to a matrix of two columns (x and y)
}
