% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{bbox_to_sfPOLYGON}
\alias{bbox_to_sfPOLYGON}
\title{Convert a 4-element bbox vector to a sf::POLYGON simple feature object}
\usage{
bbox_to_sfPOLYGON(bb = get_bb("gom"), proj_string = get_proj_string())
}
\arguments{
\item{bb}{a 4-element numeric vector of left, bottom, right, top coordinates}

\item{proj_string}{a proj4string suitable to pass to \code{sf::st_crs()}}
}
\value{
a sf:sf with a single POLYGON geometry object
}
\description{
Convert a 4-element bbox vector to a sf::POLYGON simple feature object
}
