% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{bbox_to_SpatialPoints}
\alias{bbox_to_SpatialPoints}
\title{Convert a 4-element bbox vector to SpatialPoints}
\usage{
bbox_to_SpatialPoints(bb = get_bb("gom"),
  proj_string = get_proj_string("longlat"))
}
\arguments{
\item{bb}{a 4-element numeric vector of left, bottom, right, top coordinates}

\item{proj_string}{a proj4string suitable to pass to \code{sp::CRS()}}
}
\value{
a SpatialPoints object
}
\description{
Convert a 4-element bbox vector to SpatialPoints
}

