% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{cellLayerFromIndex}
\alias{cellLayerFromIndex}
\title{Convert a index (1... ncells*nlayers) into a multilayer raster to cell and
   layer coordinates}
\usage{
cellLayerFromIndex(R, index)
}
\arguments{
\item{R}{brick or stack}

\item{index}{vector of multilayer index coordinates}
}
\value{
a tibble of index, cell, col, row, x, y, and layer
}
\description{
Convert a index (1... ncells*nlayers) into a multilayer raster to cell and
   layer coordinates
}
