% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{cellLayerFromPts}
\alias{cellLayerFromPts}
\title{Converts point location to cell number and layer number.}
\usage{
cellLayerFromPts(R, pts)
}
\arguments{
\item{R}{a multilayer Raster* object}

\item{pts}{location info for points must have at least one of the following
combinations of columns
\itemize{
    \item{'cell' if present this is returned}
    \item{'row' and 'col'}
    \item{'lon' and 'lat'}
    \item{'x' and 'y' assumed to be the same as 'lon' and 'lat'}
}}
}
\value{
vector of cell numbers
}
\description{
This function is essentially a wrapper around the \code{raster::cellFrom*()} functions.
}
