% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celmap-package.R
\docType{package}
\name{celmap}
\alias{celmap}
\alias{celmap-package}
\title{celmap: Tools working with Computational Ecology Lab map data.}
\description{
Extra functionality for working with mapped data
}

