% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearth.R
\name{celmap_bb}
\alias{celmap_bb}
\title{Various bounding boxes for analyses or forecasts}
\usage{
celmap_bb(where = c("all", "maine", "gom", "nwa", "neac", "native")[2],
  pad = list(a = c(0, 0), b = c(0.1, 0.1))[[1]], form = c("numeric",
  "extent", "sp")[1])
}
\arguments{
\item{where}{character 'native' (aka 'all'), 'nwa', 'neac' etc}
}
\value{
4 element vector of [left, right, bottom, top]
}
\description{
Note - analyses are LCC coordinates while forecasts are longlat
}

