% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{closest_available_cell}
\alias{closest_available_cell}
\title{Determine the closest non-NA pixel given [lon,lat] and a lut raster}
\usage{
closest_available_cell(
  x = dplyr::tibble(lon = c(0.1, 0.6, 0.9, 0.1), lat = c(0.3, 0.1, 0.8, 0.9)),
  lut = make_raster_lut()
)
}
\arguments{
\item{x}{the data frame with lon and lat coordinates}

\item{lut}{the raster look-up with precomputed closest non-NA cells}
}
\value{
tibble with lon and lat columns
}
\description{
Determine the closest non-NA pixel given [lon,lat] and a lut raster
}
