% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{closest_ocean_cell}
\alias{closest_ocean_cell}
\title{Determine the closest ocean pixel given [lon,lat] and a lut raster}
\usage{
closest_ocean_cell(x, lut = "/mnt/ecocast/coredata/obpg/world/lut_4km.grd")
}
\arguments{
\item{x}{the data frame with lon and lat coordinates}

\item{lut}{the raster look-up with precomputed closest ocean cells}
}
\value{
tibble with lon and lat columns
}
\description{
Determine the closest ocean pixel given [lon,lat] and a lut raster
}
