% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{extractPts}
\alias{extractPts}
\title{Extract values from a Raster* of SpatRaster object}
\usage{
extractPts(pts, x)
}
\arguments{
\item{pts}{location info for points to be extracted. Must be a data frame
 or matrix with either set of the following columns. Note that layer (or z)
 may be either a layer index number or layer names. If 'cell' is present in the
 input then [x or lon] and [y or lat] are ignored.
\itemize{
     \item{'cell' and 'layer' (or 'z')}
     \item{'row', 'col' and 'layer' (or 'z')}
     \item{'lon', 'lat' and 'layer' (or 'z')}
     \item{'x', 'y' and 'layer' (or 'z')}
 }}

\item{x}{a single or multilayer Raster* object.  If \code{x} has only one layer
then pts is passed to \code{raster::extract()} immediately.}
}
\value{
a vector of values
}
\description{
Extract values from a Raster* of SpatRaster object
}
