% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{faux_randomPoints}
\alias{faux_randomPoints}
\title{Select N random points from a faux multilayer Raster* object.}
\usage{
faux_randomPoints(R, layers = seq_len(raster::nlayers(R)), pts = NULL,
  N = 1000, M = 5)
}
\arguments{
\item{R}{a Raster* object - treated as a single layers}

\item{layers}{a vector of layer IDs, by default 1,2, ..., nlayers(R) but can be anything}

\item{pts}{xyz values for presence points [lon, lat, layer] - these
 locations are avoided when sampling. Ignored if NULL.   This must have
 either set of the following columns.  Note that layer (or z)
 must be found in the \code{layer} argument.
\itemize{
     \item{'cell' and 'layer' (or 'z')}
     \item{'row', 'col' and 'layer' (or 'z')}
     \item{'lon', 'lat' and 'layer' (or 'z')}
     \item{'x', 'y' and 'layer' (or 'z')}
 }}

\item{N}{the number of points to select}

\item{M}{numeric, a multiplier to start with a selection of random points
from which N are selected.}
}
\value{
data.frame of sampled points with the following
 \itemize{
     \item{lon}
     \item{lat}
     \item{col}
     \item{row}
     \item{cell 2D matrix index}
     \item{index 3d array index}
     \item{layer}
     \item{value - everywhere NA to match \code{layers_randomPoints()}}
 }
}
\description{
Similar to \code{layers_randomPoints()} but here the user provides a vector of layer
identifiers (indicies, names, etc) that may or may not exist in the input raster, \code{R}.
Also, NA is returned as the value for each selected point.
}
\details{
This function does not guarantee that returned locations do not coincide with
missing points.  The first layer in R is used with \code{dismo::randomPoints()}
as if the first layer is a mask so select M*N non-NA points.
}
