% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb.R
\name{get_bb}
\alias{get_bb}
\title{Various bounding boxes for analyses or forecasts}
\usage{
get_bb(where = c("all", "maine", "gom", "nwa", "neac", "native")[2],
  pad = NULL, form = c("numeric", "extent", "sp", "sf")[1], ...)
}
\arguments{
\item{where}{character 'native' (aka 'all' or 'world'), 'nwa', 'neac' etc
\itemize{
  \item{'maine' \code{-71.1, -67, 43, 47.5}}
  \item{'gom' Gulf of Maine \code{-72, -63, 39, 46}}
  \item{'nwa' Northwest Atlantic \code{-77.0, -51.5, 37.9, 56.7}}
  \item{'neac' New England Atlantic Canada \code{-74, -59.75, 41, 48.15}}
  \item{'liac' Long Island Atlantic Canada \code{-74, -59.75, 37.9, 48.15}}
  \item{'gosl' Gulf of St Lawrence \code{-67,-56.5,44.4,50.5}}
  \item{any other \code{-180, 180, -90, 90}}
}}

\item{pad}{numeric one or two element pad to add to a bounding box,
default is NULL to skip}

\item{form}{character, either numeric, extent, sf or sp}

\item{...}{further arguments - likely proj_string for sp and sf outputs}
}
\value{
4 element vector of [left, right, bottom, top], extent or SpatialPolygons object
nwa'   =   c(-77.1199677903564, -51.6967983980479,
}
\description{
Various bounding boxes for analyses or forecasts
}
