% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearth.R
\name{get_boundaries}
\alias{get_boundaries}
\title{Get a Spatial vector object by state or province name}
\usage{
get_boundaries(where = c("Maine", "New Brunswick", "Quebec",
  "New Hampshire"), what = c("boundary50m", "boundary10m")[1])
}
\arguments{
\item{where}{names the states and or provinces, or group names such as 'GOM'}

\item{what}{the dataset to quuery (boundary50m - the default - or boundary10m)}
}
\value{
a Spatial* object
}
\description{
'Quebec' can be used in lieu of "Qu\xe9bec"
}
