% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{get_crs}
\alias{get_crs}
\title{Get a projection string}
\usage{
get_crs(...)
}
\arguments{
\item{...}{arguments for \code{link{get_proj_string}}}
}
\value{
a projection string
}
\description{
Get a projection string
}
\seealso{
\code{link{get_proj_string}}
}
