% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{get_proj_string}
\alias{get_proj_string}
\title{Get one of the common projection strings}
\usage{
get_proj_string(name = c("longlat", "lcc", "utm-19", "EPSG:3857", "all")[1])
}
\arguments{
\item{name}{the name of the projection string (just one)}
}
\value{
a projection string
}
\description{
Get one of the common projection strings
}
