% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearth.R
\name{get_vectors}
\alias{get_vectors}
\title{Get a Spatial vector object by region name and vector set.}
\usage{
get_vectors(where = c("World", "Maine", "GOM", "NWA", "NEAC")[1],
  what = c("coast50m", "coast10m", "boundary50m", "boundary10m")[1], ...)
}
\arguments{
\item{where}{the name of the region}

\item{what}{the type of boundary (10m, 50m etc)}

\item{...}{further arguments for \code{celmap_bb()}}
}
\value{
a Spatial* object
}
\description{
Currently draws from Natural Earth.  Note that this clips the region
so edge effects appear - especially for boundaries.  See \code{get_boundary()}
}
