% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grd.R
\name{grd_dim}
\alias{grd_dim}
\title{Extract the dimensions from one or more raster header lists}
\usage{
grd_dim(x)
}
\arguments{
\item{x}{one or more grd lists}
}
\value{
a matrix of length(x) rows with columns nrow, ncol, ncell, nlayer
}
\description{
Extract the dimensions from one or more raster header lists
}
