% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{hexbin_points}
\alias{hexbin_points}
\title{Generate a dataset of points for use with \code{\link{st_hexbin}}}
\usage{
hexbin_points(n = 1000, bb = c(-72, -63, 39, 46))
}
\arguments{
\item{n}{the number of points to generate}

\item{bb}{4 element numeric of bounding box [left, right, bottom, top]}
}
\description{
Generate a dataset of points for use with \code{\link{st_hexbin}}
}
