% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{indexFromCellLayer}
\alias{indexFromCellLayer}
\title{Compute array indices given cell and layer}
\usage{
indexFromCellLayer(cellLayer, x)
}
\arguments{
\item{cellLayer}{2 element data frame (or tibble) with cell and layer.  If layer is
of type character we covert to integer using \code{names(x)}}

\item{x}{Raster* layer, brick or stack}
}
\value{
cellLayer with index added
}
\description{
Compute array indices given cell and layer
}
