% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{indexFromPts}
\alias{indexFromPts}
\title{Compute array indices from points - points are in the raster's CRS system}
\usage{
indexFromPts(pts, x)
}
\arguments{
\item{pts}{location info for points - see \code{vetPts()}}

\item{x}{layer, brick or stack}
}
\value{
vector of array indices as if the raster were a 3d array
}
\description{
Compute array indices from points - points are in the raster's CRS system
}
