% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{is_raster}
\alias{is_raster}
\title{Test if an object is RasterLayer, RasterStack or RasterBrick}
\usage{
is_raster(x, klass = "BasicRaster")
}
\arguments{
\item{x}{object to test}

\item{klass}{character a vector of allowed class types.  Use this to narrow
the test, say for a brick by setting klasses to 'RasterBrick'}
}
\value{
logical, TRUE if the object inherits from the specified class
}
\description{
Test if an object is RasterLayer, RasterStack or RasterBrick
}
