% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{is_sf_geometry}
\alias{is_sf_geometry}
\title{Test the geometry for inheritance}
\usage{
is_sf_geometry(x, klass = "sfc", geometry_column = "geometry")
}
\arguments{
\item{x}{sf object to test}

\item{klass}{character vector of one or more classes}

\item{geometry_column}{character the name of the geometry column, by default 'geometry'}
}
\value{
logical
}
\description{
Test the geometry for inheritance
}
