% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{is_terra}
\alias{is_terra}
\title{Test if an object is SpatRaster}
\usage{
is_terra(x)
}
\arguments{
\item{x}{object to test}
}
\value{
logical, TRUE if the object inherits from the specified class
}
\description{
Test if an object is SpatRaster
}
