% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{layers_randomPoints}
\alias{layers_randomPoints}
\title{Select N random points from a mulitlayer Raster* object}
\usage{
layers_randomPoints(R, pts = NULL, N = 1000, M = 5)
}
\arguments{
\item{R}{a multilayer Raster* object}

\item{pts}{xyz values for presence points [lon, lat, layer] - these
 locations are avoided when sampling. Ignored if NULL.   This must have
 either set of the following columns.  Note that layer (or z)
 may be either a layer index number or layer names.
\itemize{
     \item{'cell' and 'layer' (or 'z')}
     \item{'row', 'col' and 'layer' (or 'z')}
     \item{'lon', 'lat' and 'layer' (or 'z')}
     \item{'x', 'y' and 'layer' (or 'z')}
 }}

\item{N}{the number of points to select}

\item{M}{numeric, a multiplier to start with a selection of random points
from which N are selected.}
}
\value{
data.frame of sampled points with the following
 \itemize{
     \item{lon}
     \item{lat}
     \item{col}
     \item{row}
     \item{cell 2D matrix index}
     \item{index 3d array index}
     \item{layer}
     \item{value}
 }
}
\description{
This function does not guaruntee that returned locations do not coincide with
missing points.  The first layer in R is used with \code{dismo::randomPoints()}
as if the first layer is a mask so select M*N non-NA points.
Those selected points are then randomly distributed across the multiple layers
in R - which gives rise to the possibility that NA points may be selected.
}
