% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{make_dummy_points}
\alias{make_dummy_points}
\title{Make dummy points set for a given raster}
\usage{
make_dummy_points(R = make_dummy_stack(), N = 10, M = 2, na.rm = TRUE)
}
\arguments{
\item{R}{RasterLayer, Stack or Brick}

\item{N}{integer, the number of points to generate}

\item{M}{numeric, a multiplier to start with a selection of random points
from which N are selected.  Ignored if na.rm is \code{FALSE}.}

\item{na.rm}{logical, if TRUE then avoid NA cells}
}
\value{
tibble of point location info ala \code{\link{extractPts}}
}
\description{
Make dummy points set for a given raster
}
