% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{make_dummy_stack}
\alias{make_dummy_stack}
\title{Create a dummy raster stack for testing.}
\usage{
make_dummy_stack(nc = 10, nr = 10, nl = 10)
}
\arguments{
\item{nc}{integer, number of columns}

\item{nr}{integer, number of rows}

\item{nl}{integer, number of layers}
}
\value{
raster stack with some NA cells
}
\description{
Create a dummy raster stack for testing.
}
