% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{make_raster_lut}
\alias{make_raster_lut}
\title{Make a raster LUT which is essentially a lookup table that for a given
location points to the nearest non-NA value in a mask.}
\usage{
make_raster_lut(x = make_dummy_mask())
}
\arguments{
\item{x}{raster mask - where masked values are NA, unmasked values or not NA}
}
\value{
raster of cell addresses.  Where the input, R, had non-NA values the
  cell addresses point to the input cell.  Where the input had NA valued cells,
  the output cell addresses point to the nearest non-NA cells inthe input.
}
\description{
Each cell of a lut contains a cell address of the closest
cell covered by a non-NA value in the input.  Obviously, cells already
covered by non-NA will have it's original cell number.  Cells covered by
NA in the input will have a cell number of the closest non-NA cell.
}
