% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_results}
\alias{plot_results}
\title{plot_results}
\usage{
plot_results(data, results)
}
\arguments{
\item{data}{A sf object which represent the area of interest.}

\item{results}{A sf object obtained form asf_search.}
}
\value{
None. Function produces a plot
}
\description{
Plot results from asf_search.
}
\examples{

require(sf)

polygon <- cbind(c(-75.63, -75.49, -75.49, -75.63, -75.63),
                c(6.36, 6.36, 6.17, 6.17, 6.36)) \%>\%
 {st_polygon(list(.))} \%>\%
 st_sfc(crs = 4326)

results <- asf_search(data = polygon, results = 20,
 platform = 'ALOS', instrument = 'PALSAR')

plot_results(data = polygon, results = results)

}
