% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{project_table}
\alias{project_table}
\title{Transform locations in a table from one projection to another}
\usage{
project_table(x = data.frame(lon = c(-72, -63), lat = c(39, 46)),
  from_crs = "+init=epsg:4326", to_crs = "+init=epsg:32619",
  from_names = c("lon", "lat"), to_names = c("x", "y"))
}
\arguments{
\item{x}{table(data.frame or tibble) with columns specified by \code{from_names}}

\item{from_crs}{projection of the source coordinates - see \code{\link[sf]{st_crs}}}

\item{to_crs}{projection of the source coordinates - see \code{\link[sf]{st_crs}}}

\item{from_names}{string, the names of the x and y coords to project from, ['lon', 'lat'] default.}

\item{to_names}{string, the names of the x and y coords to project into, ['x', 'y'] default.
If from_names is the same as to_names then replacement occurs, otherwise two
new columns \code{to_names} are appended to the table.}
}
\value{
updated table
}
\description{
Elements of the input that are NA are returned as NA
}
\examples{
\dontrun{
x <- data.frame(lon = c(-72, -63), lat = c(39, 46))
(y <- project_table(x, to_names = c("false_easting", "false_northing")))
# lon lat false_easting false_northing
# 1 -72  39      240199.8        4321059
# 2 -63  46      964564.1        5111577
}
}
