% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{project_tibble}
\alias{project_tibble}
\title{Transform locations in a tibble or data frame from one projection to another}
\usage{
project_tibble(x, from_proj = get_proj_string("longlat"),
  to_proj = get_proj_string("lcc"), from_names = c("lon", "lat"),
  to_names = c("x", "y"))
}
\arguments{
\item{x}{tibble with columns lon and lat}

\item{from_proj}{projection of the source coordinates}

\item{to_proj}{projection of the source coordinates}

\item{from_names}{string, the names of the x and y coords to project from c('lon', 'lat') default}

\item{to_names}{string, the names of the x and y coords to project into c('x', 'y') default,
the if from_names is the same as to_names then replacement occurs}
}
\value{
updated tibble
}
\description{
Elements of the input that are NA are returned as NA
}
