% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{qadvance}
\alias{qadvance}
\title{Add an element to the front of a queue}
\usage{
qadvance(Q, x)
}
\arguments{
\item{Q}{rdeque class object}

\item{x}{an element to add to the queue, the forst on is dropped}
}
\value{
the updated queue
}
\description{
Add an element to the front of a queue
}
