% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{qavg}
\alias{qavg}
\title{Transform a queue into a raster stack and compute the mean}
\usage{
qavg(Q, ...)
}
\arguments{
\item{Q}{rdeque class object}

\item{...}{furthere arguments for \code{\link[raster]{calc}} where the
function is mean}
}
\value{
RasterLayer class
}
\description{
Transform a queue into a raster stack and compute the mean
}
