% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{qpop}
\alias{qpop}
\title{Add an element to the front of a queue}
\usage{
qpop(Q)
}
\arguments{
\item{Q}{rdeque class object}
}
\value{
the removed queue element
}
\description{
Add an element to the front of a queue
}
\seealso{
\code{\link[rstackdeque]{without_back}}
}
