% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{qpush}
\alias{qpush}
\title{Add an element to the front of a queue}
\usage{
qpush(Q, x)
}
\arguments{
\item{Q}{rdeque class object}

\item{x}{an element to add to the front of the queue}
}
\value{
the updated queue
}
\description{
Add an element to the front of a queue
}
\seealso{
\code{\link[rstackdeque]{insert_front}}
}
