% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{qstack}
\alias{qstack}
\title{Transform a queue into a raster stack}
\usage{
qstack(Q)
}
\arguments{
\item{Q}{rdeque class object}
}
\value{
RasterStack class
}
\description{
Transform a queue into a raster stack
}
