% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{randomPts}
\alias{randomPts}
\title{Generate random points within a raster object.}
\usage{
randomPts(x, n = 100, m = 2, na.rm = FALSE, pts = NULL,
  polygon = NULL)
}
\arguments{
\item{x}{Raster* layer, brick or stack}

\item{n}{numeric, the number of points to return}

\item{m}{multiplier to use when seeking to avoid missing values.
Ignored if polygon is provided.}

\item{na.rm}{logical, if TRUE then avoid cells with missing values.}

\item{pts}{a table of points to avoid. If na.rm is TRUE that is handled first,
then point avoidance is handled.  Ignored if NULL.}

\item{polygon}{polygon (sfc or SpatialPolygon) that describes the region to select
points from.  Ignored if NULL.}
}
\value{
a table of locations with values.  Note that it is possible
        to filter the available pool of candidate cells to something
        less than the requested sample, n.  In such cases some
        samples may be replicates.
}
\description{
Providing a polygon to limit the search area can speed up the search
if the polygon is small relative to the area of a layer.
}
