% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasf-package.R
\docType{package}
\name{rasf}
\alias{rasf}
\title{rasf: Tools for working with rasters and simple features.}
\description{
Extra functionality for working with raster (raster and terra packages) and sf data.
}
