% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{raster_crop}
\alias{raster_crop}
\title{Crop a raster - a wrapper around \code{raster::crop} or \code{raster::crop}}
\usage{
raster_crop(x, ...)
}
\arguments{
\item{x}{RasterLayer or SpatRaster}

\item{...}{other arguments for \code{raster::crop} or \code{raster::crop}}
}
\value{
Raster or SpatRaster object
}
\description{
Crop a raster - a wrapper around \code{raster::crop} or \code{raster::crop}
}
