% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{raster_dim}
\alias{raster_dim}
\title{Assemble a vector of raster dimension}
\usage{
raster_dim(x)
}
\arguments{
\item{x}{Raster object}
}
\value{
numeric vector of
\itemize{
  \item{ncol the number of columns}
  \item{nrow the number of rows}
  \item{ncell the number of cells (pixels) in each layer (ncol * nrow)}
  \item{nlayer the number of layers (bands)}
  \item{nindex the number of indices (ncell*nlayer)}
}
}
\description{
Assemble a vector of raster dimension
}
\examples{
\dontrun{
slogo <- stack(system.file("external/rlogo.grd", package="raster"))
raster_dim(slog)
}
}
