% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{raster_fileinfo}
\alias{raster_fileinfo}
\title{Retrieve metadata about one or more raster files}
\usage{
raster_fileinfo(x)
}
\arguments{
\item{x}{character, one or more filenames (.tiff, .grd, etc.)}
}
\value{
table (tibble) of GDAL metadata
\itemize{
  \item{rows integer }
  \item{columns integer }
  \item{bands integer }
  \item{ll.x numeric }
  \item{ll.y numeric }
  \item{res.x numeric }
  \item{res.y numeric }
  \item{oblique.x numeric }
  \item{oblique.y numeric }
  \item{crs character }
}
}
\description{
This is a wrapper around \code{\link[rgdal]{GDALinfo}}
}
