% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{raster_rotate}
\alias{raster_rotate}
\title{A wrapper around \code{\link[raster]{rotate}} that provides rotation from
[-180,180] to [0,360].}
\usage{
raster_rotate(x, inverse = FALSE, adjust_origin = TRUE, filename = "", ...)
}
\arguments{
\item{x}{raster, a raster object with global coverage}

\item{inverse}{logical if FALSE then pass all of the arguments to
\code{\link[raster]{rotate}}.  If FALSE then transform from [-180,180]
to [0,360]}

\item{adjust_origin}{logical, if TRUE then make minor adjustment to origin.
Ignored if \code{inverse} is FALSE.}

\item{filename}{character see \code{\link[raster]{merge}}}

\item{...}{optional arguments for \code{\link[raster]{merge}}}
}
\value{
raster object
}
\description{
A wrapper around \code{\link[raster]{rotate}} that provides rotation from
[-180,180] to [0,360].
}
