% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{raster_shape}
\alias{raster_shape}
\title{Assemble a vector of raster dimension}
\usage{
raster_shape(x)
}
\arguments{
\item{x}{Raster object}
}
\value{
numeric vector of nc = ncols, nr = nrows, n = ncells (2d),
    nl = nlayers, N = total number of cells (3d)
}
\description{
Assemble a vector of raster dimension
}
