% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{raster_shift}
\alias{raster_shift}
\title{Shift a raster - a wrapper around \code{raster::shift} or \code{raster::shift}}
\usage{
raster_shift(x, ...)
}
\arguments{
\item{x}{RasterLayer or SpatRaster}

\item{...}{other arguments for \code{raster::shift} or \code{raster::shift}}
}
\value{
Raster or SpatRaster object
}
\description{
Shift a raster - a wrapper around \code{raster::shift} or \code{raster::shift}
}
