% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkt.R
\name{raster_use_wkt}
\alias{raster_use_wkt}
\title{Test if we need WKT or proj4string for raster package}
\usage{
raster_use_wkt(threshold = "4")
}
\arguments{
\item{threshold}{character, the version that determines if wkt is required}
}
\value{
logical TRUE if WKT is required
}
\description{
Test if we need WKT or proj4string for raster package
}
