% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{sf_geometry}
\alias{sf_geometry}
\title{Retrieve the class of the geometry}
\usage{
sf_geometry(x, geometry_column = "geometry")
}
\arguments{
\item{x}{sf object}

\item{geometry_column}{character the name of the geometry column, by default 'geometry'}
}
\value{
character
}
\description{
Retrieve the class of the geometry
}
