% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkt.R
\name{sf_use_wkt}
\alias{sf_use_wkt}
\title{Test if we need WKT or proj4string for sf package}
\usage{
sf_use_wkt(name = "proj.4", threshold = "6")
}
\arguments{
\item{name}{character, the name of the sf dependency to test}

\item{threshold}{character, the dependency version that determines if wkt is required}
}
\value{
logical TRUE if WKT is required
}
\description{
Test if we need WKT or proj4string for sf package
}
