% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexbin.R
\name{st_hexbin_utm}
\alias{st_hexbin_utm}
\title{Retrieve of set of hex-binned polygons for the given xyz data}
\usage{
st_hexbin_utm(x, y, z, fun = c("mean", "count", "median")[2],
  crs = "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0",
  utm_crs = "+proj=utm +zone=19 ellps=WGS84", na.rm = TRUE, ...)
}
\arguments{
\item{x}{a vector of lon-locations}

\item{y}{a vector of lat-locations}

\item{z}{a vector of values associated with each location}

\item{fun}{character the function used to process the cell values
currently only 'mean', 'median', or 'count' (default)}

\item{crs}{the projection string for final data}

\item{utm_crs}{intermediary UTM projection string}

\item{na.rm}{logical passed through to user's selected function}

\item{...}{further arguments for \code{\link[hexbin]{hexbin}}}
}
\value{
a POLYGON sf data frame
}
\description{
Retrieve of set of hex-binned polygons for the given xyz data
}
\seealso{
\href{https://www.rdocumentation.org/packages/hexbin/versions/1.29.0/topics/hexbin}{hexbin docs}

\href{https://hexnet.org/content/hexagonal-geometry}{nice overview}
}
