% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb.R
\name{to360}
\alias{to360}
\title{Convert [-180,180] longitudes to [0, 360]}
\usage{
to360(x)
}
\arguments{
\item{x}{numeric vector, no check is done for being within [0,3 60] range}
}
\value{
numeric vector
}
\description{
Convert [-180,180] longitudes to [0, 360]
}
\seealso{
\url{https://gis.stackexchange.com/questions/201789/verifying-formula-that-will-convert-longitude-0-360-to-180-to-180/201793}
}
