% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkt.R
\name{use_wkt}
\alias{use_wkt}
\title{Test if WKT is required for specifying coordinate reference}
\usage{
use_wkt(package = c("sf", "raster")[1], ...)
}
\arguments{
\item{package}{character, the name of the package to test}

\item{...}{arguments for \code{\link{sf_use_wkt}} or \code{\link{raster_use_wkt}}}
}
\value{
logical TRUE if WKT is required
}
\description{
Test if WKT is required for specifying coordinate reference
}
