% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{vet_pts}
\alias{vet_pts}
\title{Vet points to ensure uniform input}
\usage{
vet_pts(pts)
}
\arguments{
\item{pts}{a matrix, data.frame or tibble}
}
\value{
a tibble of x, y and layer
}
\description{
The following are synonyms and definitions
\itemize{
   \item{x: x, lon}
   \item{y: y, lat}
   \item{layer: layer, z, name}
}
}
