% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{vetPts}
\alias{vetPts}
\title{Vet points to ensure uniform input}
\usage{
vetPts(pts)
}
\arguments{
\item{pts}{a (column-named) matrix, data.frame or tibble}
}
\value{
a tibble of x, y and layer
}
\description{
The following are synonyms and definitions.  Note that layer is not required,
if not present then it is assigned the value 1 for all points.
\itemize{
   \item{x: x, lon}
   \item{y: y, lat}
   \item{layer: layer, z, name}
}
}
