% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{volcano_points}
\alias{volcano_points}
\title{Generate a table of random points in a raster stack (or layer)}
\usage{
volcano_points(x = volcano_stack(), n = 100, ...)
}
\arguments{
\item{x}{RasterLayer or RasterStack}

\item{n}{numeric, the numer of points to select}

\item{...}{further arguments for \link{randomPts}}
}
\value{
tibble of locations with values
}
\description{
Generate a table of random points in a raster stack (or layer)
}
