% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{volcano_polygon}
\alias{volcano_polygon}
\title{Generate a polygon for the volcano raster.}
\usage{
volcano_polygon()
}
\value{
simple feature geometry for a POLYGON
}
\description{
Generate a polygon for the volcano raster.
}
\examples{
\dontrun{
library(dplyr)
x <- volcano_stack(indexed = TRUE)
p <- volcano_polygon()
pts <- randomPts(x, polygon = p)
par(mfrow = c(1,3))
for (i in seq_len(3)){
  plot(x[[i]], main = paste("Layer", i))
  plot(p, add = TRUE)
  with(pts \%>\% dplyr::filter(layer == i), points(x, y))
}
}
}
