% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{volcano_stack}
\alias{volcano_stack}
\title{Retrieve a raster stack of Auckland’s Maunga Whau volcano topography.
Layers 2 to nlayers slightly altered from the original as per
\code{\link{volcano_raster}}}
\usage{
volcano_stack(nlayers = 3, indexed = FALSE)
}
\arguments{
\item{nlayers}{numeric, the number of layers for the stack}

\item{indexed}{logical, if TRUE then assign 1,2,3,... as cell values}
}
\value{
RasterStack
}
\description{
Retrieve a raster stack of Auckland’s Maunga Whau volcano topography.
Layers 2 to nlayers slightly altered from the original as per
\code{\link{volcano_raster}}
}
\seealso{
\code{\link{volcano_raster}}
}
