% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{xyCellLayerFromIndex}
\alias{xyCellLayerFromIndex}
\title{Convert a index (1, 2, 3,... ncells*nlayers) into a multilayer raster to row, col, cell and
   layer coordinates}
\usage{
xyCellLayerFromIndex(index, x)
}
\arguments{
\item{index}{vector of multilayer index coordinates}

\item{x}{Raster* layer, brick or stack}
}
\value{
a tibble of index, cell, col, row, x, y, and layer
}
\description{
Convert a index (1, 2, 3,... ncells*nlayers) into a multilayer raster to row, col, cell and
   layer coordinates
}
