% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asf_search.R
\name{asf_search}
\alias{asf_search}
\title{asf_search}
\usage{
asf_search(data, results = 25, platform, instrument = NULL, start = NULL)
}
\arguments{
\item{data}{A sf object which represent the area where search the selected
ASF product.}

\item{results}{An integer, specifying the maximum number of results returned
from search.}

\item{platform}{A string specifying the ASF product for search. Available
products are:

ALOS, JERS-1, RADARSAT-1, Sentinel, Sentinel-1, Sentinel-1A, Sentinel-1B, SMAP.}

\item{instrument}{Default NULL. For some platforms, such as ALOS, there are
multiple instruments to choose from.}

\item{start}{Start Date of search. Default NULL.}
}
\value{
A data frame with available products for the selected area.
}
\description{
Retrieve a collection with available ASF products.
}
\examples{

require(sf)
require(readr)

polygon <- cbind(c(-75.63, -75.49, -75.49, -75.63, -75.63),
                c(6.36, 6.36, 6.17, 6.17, 6.36)) \%>\%
 {st_polygon(list(.))} \%>\%
 st_sfc(crs = 4326)

asf_search(data = polygon, results = 5,
           platform = 'ALOS', instrument = 'PALSAR')

}
\seealso{
https://docs.asf.alaska.edu/api/keywords/
}
