% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb.R
\name{bb_to_polygon}
\alias{bb_to_polygon}
\title{Convert a 4-element bounding box vector to a polygon simple feature object}
\usage{
bb_to_polygon(bb = c(-72, -63, 39, 46), crs = "+init=epsg:4326")
}
\arguments{
\item{bb}{a 4-element numeric vector of left, right, bottom and top coordinates}

\item{crs}{character, the coordinate reference system}
}
\value{
a simple feature with a POLYGON
}
\description{
Convert a 4-element bounding box vector to a polygon simple feature object
}
