% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{indexFromPts}
\alias{indexFromPts}
\title{Compute array indices from points}
\usage{
indexFromPts(R, pts)
}
\arguments{
\item{R}{brick or stack}

\item{pts}{location info for points - see \code{vet_pts()}}
}
\value{
vector of array indices as if the raster were a 3d array
}
\description{
Compute array indices from points
}
