% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{make_raster_lut}
\alias{make_raster_lut}
\title{Make a raster LUT which is essentially a lookup table that for a given
location points to the nearest non-mask value in a mask.}
\usage{
make_raster_lut(x = make_dummy_mask(), mask_value = NA)
}
\arguments{
\item{x}{raster mask}

\item{mask_value}{numeric the value of masked areas, by default NA}
}
\value{
raster of cell addresses.  Where the input, R, had non-mask values the
  cell addresses point to the input cell.  Where the input had mask-valued cells,
  the output cell addresses point to the nearest non-mask cells in the input.
}
\description{
Each cell of a lut contains a cell address of the closest
cell covered by a non-mask value in the input.  Obviously, cells already
covered by non-mask will have it's original cell number.  Cells covered by
\code{mask_value} in the input will have a cell number of the closest
non-mask cell.
}
\details{
Note this assigns the closest cell value which is a 1-d index into 2-d space.
}
