% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{read_grd}
\alias{read_grd}
\title{Read the header portion of a raster file}
\usage{

  read_grd(filename = "/mnt/ecocast/coredata/obpg/gom/DAY/chlor_a/A2014365.L3m_DAY_CHL_chlor_a_4km.grd")
}
\arguments{
\item{filename}{on or more filenames}
}
\value{
one or more grd lists
\itemize{
     \item{dim vector of nrows, ncols, ncells, nlayers}
     \item{extent vector of xmin, xmax, ymin, ymax}
     \item{projection  string}
 }
}
\description{
Read the header portion of a raster file
}
