% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexbin.R
\name{st_hexbin}
\alias{st_hexbin}
\title{Retrieve of set of hex-binned polygons for the given xyz data}
\usage{
st_hexbin(x, y, z, fun = c("mean", "count", "median")[2],
  crs = ifelse(use_wkt("sf"), "epsg:4326",
  "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0"),
  na.rm = TRUE, ...)
}
\arguments{
\item{x}{a vector of x-locations}

\item{y}{a vector of y-locations}

\item{z}{a vector of values associated with each location}

\item{fun}{character the function used to process the cell values
currently only 'mean', 'median', or 'count' (default)}

\item{crs}{the projection string}

\item{na.rm}{logical passed through to user's selected function}

\item{...}{further arguments for \code{\link[hexbin]{hexbin}}}
}
\value{
a POLYGON sf data frame
}
\description{
Retrieve of set of hex-binned polygons for the given xyz data
}
\examples{
\dontrun{
xyz <- hexbin_points()
 hb <- st_hexbin(xyz$x, xyz$y, xyz$z)
 hb <- dplyr::rename(hb, count = value)
 pts <- sf::st_as_sf(xyz,
                     coords = c("x", "y"),
                     crs = "+init=epsg:4326",
                     agr = "identity")
 plot(hb, asp = 1, reset = FALSE, axes = TRUE)
 plot(pts, col = 'green', add = TRUE)
}
}
\seealso{
\href{https://www.rdocumentation.org/packages/hexbin/versions/1.29.0/topics/hexbin}{hexbin docs}

\href{https://hexnet.org/content/hexagonal-geometry}{nice overview}
}
