% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{volcano_raster}
\alias{volcano_raster}
\title{Retrieve a raster of Auckland’s Maunga Whau volcano topography.  Adapted from
USGS-R inlmisc package.}
\usage{
volcano_raster(indexed = FALSE, klass = c("RasterLayer", "SpatRaster")[2])
}
\arguments{
\item{indexed}{logical, if TRUE then assign 1,2,3,... as cell values}

\item{klass}{character, one of \code{RasterLayer} of \code{SpatRaster}}
}
\value{
RasterLayer or SpatRaster
}
\description{
Retrieve a raster of Auckland’s Maunga Whau volcano topography.  Adapted from
USGS-R inlmisc package.
}
\seealso{
\url{https://waterdata.usgs.gov/blog/inlmiscmaps/}

\url{https://CRAN.R-project.org/package=inlmisc}
}
